Rely on repo's "make docker" instead of ko.
We're going to use kaniko for reproducible builds instead of ko.  This
commit relies on the target repository's "make docker" for reproducible
builds; which is going to invoke kaniko.

This commit also adds "--no-cache" to the builder image, which compiles
a docker image to an EIF.  This makes sure that we're always going to
have the latest version of nitro-cli.